feat: support amp

+++ models/imagenet/utils/misc.py
@@ -25,7 +25,8 @@ def accuracy(output, target, topk=(1,), raw=False):
         maxk = max(topk)
 
         _, pred = output.topk(maxk, 1, True, True)
-        pred = pred.t()
+        pred = pred.t().to(target.dtype)
+
         correct = pred.eq(target.view(1, -1).expand_as(pred))
 
         res = []
